requirements.txt
# Add project dependencies here
# Example:
# requests>=2.25.0
# numpy>=1.21.0
requests>=2.25.0
beautifulsoup4>=4.9.0
lxml>=4.6.0
anthropic>=0.3.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0 